requirements.txt
Django==5.1.3
drf-yasg==1.21.8
django-filter~=24.3
djangorestframework~=3.15.2
django-simple-history~=3.7.0
django-import-export~=4.3.3